КРАД на пути к совершенству

+++ server/frontend/controllers/task/Delete.php
@@ -4,6 +4,9 @@ namespace frontend\controllers\task;
 
 use frontend\components\RestAction;
 use common\models\TaskRecord;
+use common\models\IssueRecord;
+use common\models\ProjectRecord;
+use common\models\CustomerRecord;
 use Yii;
 
 class Delete extends RestAction
@@ -17,6 +20,17 @@ class Delete extends RestAction
 			return;
 		}
 		
+		$ownerId = CustomerRecord::findOne(
+			ProjectRecord::findOne(
+				IssueRecord::findOne($task->issue_id)->project_id
+			)->customer_id
+		)->user_id;
+
+		if ($ownerId !== Yii::$app->user->identity->id) {
+			Yii::$app->getResponse()->setStatusCode(403);
+			return;
+		}
+
 		if (false === $task->delete()) {
 			throw new Exception('Detetion of Task was unsuccessfull');
 			return false;
